macroification of fe/scalar/curve with fiat
