+++ src/curve/fiat/mod.rs
@@ -10,4 +10,5 @@ pub mod p521_scalar_64;
 pub mod secp256k1_64;
 pub mod secp256k1_scalar_64;
 
+mod curve_macros;
 mod field_macros;
