+++ src/curve/sec2/mod.rs
@@ -2,5 +2,7 @@ mod bigint;
 
 #[cfg(feature = "p256k1")]
 pub mod p256k1;
+#[cfg(feature = "p256r1")]
+pub mod p256r1;
 
 pub use bigint::*;
